implemented customizable state-based deserialization
allows user to define three functions in an 'options' object:
    - getStartState
    - processKeyValue
    - processComment
This options object can then be passed to the deserializer when initialized.  If the object is not provided, default functions will be used, allowing for complete backward compatability.

The advantage is that the user can override the default behavior, allowing them to customize how the file is parsed.

For example, if comments following the form:  "# translate: true" are to mark a block of text which should be translated, the user can specify three basic functions which will watch for these comments and change the parsing behavior accordingly (see the added test suite in test/properties.js).